Clean up integration test cases (#884)
* Clean up integration test cases
