--- test_shard/fake_test_app/integration_test/musiccontentcenter_fake_test.dart
@@ -1,9 +1,0 @@
-import 'package:integration_test/integration_test.dart';
-
-import 'generated/musiccontentcenter_fake_test.generated.dart' as generated;
-
-void main() {
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-
-  generated.musicContentCenterSmokeTestCases();
-}
