+++ test_shard/fake_test_app/integration_test/testcases/localspatialaudioengine_testcases.dart
@@ -4,14 +4,12 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:iris_tester/iris_tester.dart';
 import 'package:iris_method_channel/iris_method_channel.dart';
-import 'generated/localspatialaudioengine_fake_test.generated.dart'
+import '../generated/localspatialaudioengine_fake_test.generated.dart'
     as generated;
-import 'generated/basespatialaudioengine_fake_test.generated.dart'
+import '../generated/basespatialaudioengine_fake_test.generated.dart'
     as basespatialaudioengine;
 
-void main() {
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-
+void testCases() {
   generated.localSpatialAudioEngineSmokeTestCases();
   basespatialaudioengine.localSpatialAudioEngineSmokeTestCases();
