--- test_shard/integration_test_app/integration_test/mediaplayercachemanager_smoke_test.dart
@@ -1,10 +1,0 @@
-import 'package:integration_test/integration_test.dart';
-
-import 'generated/mediaplayercachemanager_smoke_test.generated.dart'
-    as generated;
-
-void main() {
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-
-  generated.mediaPlayerCacheManagerSmokeTestCases();
-}
