+++ test_shard/integration_test_app/integration_test/testcases/agora_video_view_testcases.dart
@@ -30,9 +30,7 @@ class _RenderViewWidgetState extends State<_RenderViewWidget> {
   }
 }
 
-void main() {
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
-
+void testCases() {
   testWidgets('Show Local AgoraVideoView pressure test',
       (WidgetTester tester) async {
     app.main();
