+++ test_shard/integration_test_app/integration_test/testcases/mediaplayer_smoke_test_testcases.dart
@@ -6,8 +6,7 @@ import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:integration_test_app/main.dart' as app;
 
-void main() {
-  IntegrationTestWidgetsFlutterBinding.ensureInitialized();
+void testCases() {
 
   testWidgets(
     'registerAudioFrameObserver smoke test',
